refactor: Extract RunCommand into a separate file
